cleanup Game.update, add Shortcuts.lua
