+++ input.lua
@@ -229,78 +229,6 @@ function love.keypressed(key, scancode, rep)
     return love.keyboard.isDown("lalt") or love.keyboard.isDown("ralt")
   end
 
-  local function ctrlDown()
-    return love.keyboard.isDown("rctrl") or love.keyboard.isDown("lctrl")
-  end
-
-  local function shiftDown()
-    return love.keyboard.isDown("lshift") or love.keyboard.isDown("rshift")
-  end
-
-  local function smashDown()
-    return ctrlDown() and altDown() and shiftDown()
-  end
-
-  local function handleCopy()
-    if key == "c" and not rep and ctrlDown() then
-      local stacks = {}
-      if GAME.match.P1 then
-        stacks["P1"] = GAME.match.P1:toPuzzleInfo()
-      end
-      if GAME.match.P2 then
-        stacks["P2"] = GAME.match.P2:toPuzzleInfo()
-      end
-      if tableUtils.length(stacks) > 0 then
-        love.system.setClipboardText(json.encode(stacks))
-        return true
-      end
-    end
-  end
-
-  local function handleDumpAttackPattern()
-    if (key == "1" or key == "2") and not rep and ctrlDown() then
-      local stack = GAME.match.P1
-
-      if key == "2" then
-        stack = GAME.match.P2
-      end
-
-      if stack then
-        local data, state = stack:getAttackPatternData()
-        saveJSONToPath(data, state, "dumpAttackPattern.json")
-        return true
-      end
-    end
-  end
-
-  local function modifyWinCounts()
-    if GAME.battleRoom and altDown() then
-      if key == "1" then -- Add to P1's win count
-        GAME.battleRoom.modifiedWinCounts[1] = math.max(0, GAME.battleRoom.modifiedWinCounts[1] + 1)
-      end
-      if key == "2" then -- Subtract from P1's win count
-        GAME.battleRoom.modifiedWinCounts[1] = math.max(0, GAME.battleRoom.modifiedWinCounts[1] - 1)
-      end
-      if key == "3" then -- Add to P2's win count
-        GAME.battleRoom.modifiedWinCounts[2] = math.max(0, GAME.battleRoom.modifiedWinCounts[2] + 1)
-      end
-      if key == "4" then -- Subtract from P2's win count
-        GAME.battleRoom.modifiedWinCounts[2] = math.max(0, GAME.battleRoom.modifiedWinCounts[2] - 1)
-      end
-      if key == "5" then -- Reset key
-        GAME.battleRoom.modifiedWinCounts[1] = 0
-        GAME.battleRoom.modifiedWinCounts[2] = 0
-      end
-    end
-  end
-
-  -- TODO: Implement these in Game as they access game data
-  if handleCopy() or
-     handleDumpAttackPattern() or
-     modifyWinCounts() then
-    return
-  end
-
   if not rep then
     keys[key] = 0
   end
